+++ common/src/main/java/net/pantasystem/milktea/common/paginator/EntityConverter.kt
@@ -1,0 +1,9 @@
+package net.pantasystem.milktea.common.paginator
+
+
+/**
+ * DTOをEntityに変換し共通のDataStoreにEntityを追加するためのInterface
+ */
+interface EntityConverter<DTO, E> {
+    suspend fun convertAll(list: List<DTO>) : List<E>
+}
