+++ common/src/main/java/net/pantasystem/milktea/common/paginator/FuturePaginator.kt
@@ -1,0 +1,5 @@
+package net.pantasystem.milktea.common.paginator
+
+interface FuturePaginator {
+    suspend fun loadFuture()
+}
