+++ common/src/main/java/net/pantasystem/milktea/common/paginator/IdGetter.kt
@@ -1,0 +1,10 @@
+package net.pantasystem.milktea.common.paginator
+
+/**
+ * ページネーションするためのIdを取得するためのインターフェース
+ * IdはAPIやデータベースにアクセスするものと直接的に対応する
+ */
+interface IdGetter<T> {
+    suspend fun getUntilId(): T?
+    suspend fun getSinceId(): T?
+}
