+++ common/src/main/java/net/pantasystem/milktea/common/paginator/PreviousLoader.kt
@@ -1,0 +1,10 @@
+package net.pantasystem.milktea.common.paginator
+
+import retrofit2.Response
+
+/**
+ * 過去のページを取得するためのAPIのアダプター
+ */
+interface PreviousLoader<DTO> {
+    suspend fun loadPrevious(): Response<List<DTO>>
+}
