+++ common/src/main/java/net/pantasystem/milktea/common/paginator/PreviousPaginator.kt
@@ -1,0 +1,5 @@
+package net.pantasystem.milktea.common.paginator
+
+interface PreviousPaginator {
+    suspend fun loadPrevious()
+}
