+++ common/src/main/java/net/pantasystem/milktea/common/paginator/StateLocker.kt
@@ -1,0 +1,10 @@
+package net.pantasystem.milktea.common.paginator
+
+import kotlinx.coroutines.sync.Mutex
+
+/**
+ * 状態をロックすることのできることを表すインターフェース
+ */
+interface StateLocker {
+    val mutex: Mutex
+}
