+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/gallery/MediatorGalleryPostPaginator.kt
@@ -2,15 +2,14 @@ package net.pantasystem.milktea.data.infrastructure.gallery
 
 import net.pantasystem.milktea.data.api.misskey.MisskeyAPIProvider
 import net.pantasystem.milktea.common.Encryption
-import net.pantasystem.milktea.common.FuturePagingController
-import net.pantasystem.milktea.common.PreviousPagingController
-import net.pantasystem.milktea.common.StateLocker
-
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import net.pantasystem.milktea.common.PageableState
 import net.pantasystem.milktea.common.StateContent
+import net.pantasystem.milktea.common.paginator.FuturePagingController
+import net.pantasystem.milktea.common.paginator.PreviousPagingController
+import net.pantasystem.milktea.common.paginator.StateLocker
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.account.page.Pageable
 import net.pantasystem.milktea.model.drive.FilePropertyDataSource
