+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/gallery/liked_gallery_posts_paginator.kt
@@ -10,6 +10,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.map
 import net.pantasystem.milktea.common.*
+import net.pantasystem.milktea.common.paginator.*
 import net.pantasystem.milktea.data.infrastructure.*
 import net.pantasystem.milktea.model.account.Account
 import net.pantasystem.milktea.model.drive.FilePropertyDataSource
